Split Chaincode command into multiple files
