+++ peer/chaincode/query.go
@@ -1,0 +1,44 @@
+/*
+Copyright IBM Corp. 2016 All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+		 http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package chaincode
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+func queryCmd() *cobra.Command {
+	chaincodeQueryCmd.Flags().BoolVarP(&chaincodeQueryRaw, "raw", "r", false, "If true, output the query value as raw bytes, otherwise format as a printable string")
+	chaincodeQueryCmd.Flags().BoolVarP(&chaincodeQueryHex, "hex", "x", false, "If true, output the query value byte array in hexadecimal. Incompatible with --raw")
+
+	return chaincodeQueryCmd
+}
+
+var chaincodeQueryCmd = &cobra.Command{
+	Use:       "query",
+	Short:     fmt.Sprintf("Query using the specified %s.", chainFuncName),
+	Long:      fmt.Sprintf(`Query using the specified %s.`, chainFuncName),
+	ValidArgs: []string{"1"},
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return chaincodeQuery(cmd, args)
+	},
+}
+
+func chaincodeQuery(cmd *cobra.Command, args []string) error {
+	return chaincodeInvokeOrQuery(cmd, args, false)
+}
